+++ changelog.md
@@ -2,6 +2,12 @@
 
 ## 实验阶段
 
+### WIP
+
+- [add] 增加了 `ZhihuClient.from_url` 方法，传入合法的知乎网址，就能生成对应的对象
+- [add] 给`BaseGenerator`增加了 `add_params`和`set_params` 方法
+- [change] 改变了好多内部类名和变量名，不过对外部接口没有影响
+
 ### 0.0.2 - 2016.04.07
 
 - [fix] 修复错误的 BASE_HTML_HEADER 值。原值会导致 html 文件在 Firefox 中打开时，由于没有编码信息显示而不正确的问题。